+++ src/util/commonUtil.ts
@@ -74,3 +74,9 @@ export function injectCanvas(): CanvasRenderingContext2D[] {
 
 	return contexts;
 }
+
+export function setImg(element: HTMLElement | null, image: string): void {
+	if (element instanceof HTMLImageElement) {
+		element.src = image;
+	}
+}